requirements.txt
# attendance-api-new/requirements.txt
flask==2.2.5
flask-cors==4.0.0
tensorflow==2.10.0
keras==2.10.0
opencv-python==4.6.0.66
mtcnn==0.1.1
scikit-learn==1.1.3
numpy==1.21.6
pandas==1.5.3
pillow==9.5.0
mysql-connector-python==8.0.33
python-dotenv==1.0.0
requests==2.31.0
keras-facenet==0.3.2
h5py==3.7.0